requirements.txt
streamlit==1.33.0
fastapi==0.109.0
langchain-core==0.2.0
langchain-google-genai==0.1.0
google-api-python-client==2.120.0
python-dotenv==1.0.0
uvicorn==0.27.0
requests==2.31.0